Remove ship-specific wires, only broadcast updates
If we have the need to send updates to a specific ship, we can do add that
later. As it stands, there are often (harmless, but annoying) "subscribe not
unique" errors because of how subscription to the provider are being handled. I
guess that's what I get for cargo-culting based on what's being done in the
btc-provider.