Implement AddHoldInvoice RPC
It was thought we could use the HTLCInterceptor feature of LND to forward HTLCs
to urbit. However, HTLCs get rejected before the interceptor if there is no
outgoing channel, and since LND doesn't know about urbit channels, we'll never
get that HTLC.

HODL invoices provide most of what we want: the provider generates a payment
request on behalf of nodes it's providing for. It does not know the preimage,
and must get it from the payee node by entering into an HTLC with it, at which
point it is passed back to LND for settlement.